+++ src/app/(service)/layout.tsx
@@ -36,7 +36,7 @@ export default function ServiceLayout({
   }
 
   return (
-    <html lang="en">
+    <html lang="ko">
       <head>{GTM_ID && <GoogleTagManager gtmId={GTM_ID} />}</head>
       <body className={clsx(pretendard.className, 'h-screen w-screen')}>
         <MainProvider>
